legacy_streamlit/requirements.txt
streamlit==1.54.0
psycopg2-binary==2.9.11
pandas==2.3.3
plotly==6.6.0
